se termina de enlazar los handlers del 90% de los campos
