+++ src/components/DatosGenerales/OtrasObligaciones.js
@@ -26,7 +26,8 @@ import MenuItem from "@material-ui/core/MenuItem";
 // import ListItemText from "@material-ui/core/ListItemText";
 /*Multiselect*/
 
-// import FormControlLabel from "@material-ui/core/FormControlLabel";
+import FormControlLabel from "@material-ui/core/FormControlLabel";
+import Checkbox from "@material-ui/core/Checkbox";
 
 // data picker
 // import moment from "moment";
@@ -340,15 +341,19 @@ function SimpleCard(props) {
               onChange={handleChange("porcentaje_obligacion_titular")}
               margin="normal"
             />
-          </Grid>
-          <Grid item xs={3}>
-            <TextField
-              id="grado"
+          </Grid>          
+          <Grid item xs={4}>
+            <br />
+            <FormControlLabel
+              control={
+                <Checkbox
+                  checked={data.datos_otras_obligaciones.garantia}
+                  value="true"
+                  onChange={handleChange("garantia")}
+                  color="primary"
+                />
+              }
               label="¿Se otorgó garantía?"
-              className={classes.textField}
-              value={data.datos_otras_obligaciones.garantia}
-              onChange={handleChange("garantia")}
-              margin="normal"
             />
           </Grid>
